Remove old tree structure
