+++ packages/ndla-ui/src/TreeStructure/helperFunctions.ts
@@ -7,7 +7,6 @@
  */
 
 import { IFolder } from "@ndla/types-backend/myndla-api";
-import { TreeStructureType } from "./types";
 
 export const flattenFolders = (folders: IFolder[], openFolders?: string[]): IFolder[] => {
   return folders.reduce((acc, { subfolders, id, ...rest }) => {
@@ -16,8 +15,4 @@ export const flattenFolders = (folders: IFolder[], openFolders?: string[]): IFol
     }
     return acc.concat({ subfolders, id, ...rest }, flattenFolders(subfolders, openFolders));
   }, [] as IFolder[]);
-};
-
-export const treestructureId = (type: TreeStructureType, modifier: string) => {
-  return `${type}-treestructure-${modifier}`;
 };
