+++ packages/ndla-ui/src/TreeStructure/types.ts
@@ -7,9 +7,7 @@
  */
 
 import { ReactNode } from "react";
-import { IFolder, IResource } from "@ndla/types-backend/myndla-api";
-
-export type TreeStructureType = "navigation" | "picker";
+import { IFolder } from "@ndla/types-backend/myndla-api";
 
 export type OnCreatedFunc = (folder: IFolder | undefined) => void;
 
@@ -22,22 +20,3 @@ export type NewFolderInputFunc = ({
   parentId: string;
   onCreate: OnCreatedFunc;
 }) => ReactNode;
-
-export interface CommonTreeStructureProps {
-  loading?: boolean;
-  targetResource?: IResource;
-  type: TreeStructureType;
-}
-
-export interface CommonFolderItemsProps extends CommonTreeStructureProps {
-  focusedFolder?: IFolder;
-  level: number;
-  maxLevel: number;
-  selectedFolder?: IFolder;
-  onCloseFolder: (id: string) => void;
-  onOpenFolder: (id: string) => void;
-  setFocusedFolder: (folder: IFolder) => void;
-  setSelectedFolder: (folder: IFolder) => void;
-  visibleFolders: IFolder[];
-  closeTree: () => void;
-}
